requirements.txt
Flask==2.0.2
Flask-MySQLdb==0.2.0
Flask-Cors==3.0.10
Werkzeug==2.0.2
pyjwt==2.3.0
mysqlclient==2.0.3
gunicorn==20.1.0
Flask-Mail==0.9.1




